Fix clippy warning and linker error in reference build
- Remove unnecessary `as f32` cast on quad_decimate (already f32)
- Add missing `common/image_u8_parallel.c` to reference C build sources,
  fixing undefined symbol `image_u8_gaussian_blur_parallel` when building
  with --all-features

+++ apriltag-bench/src/main.rs
@@ -401,7 +401,7 @@ fn cmd_compare(category: Option<String>, scenario: Option<String>, format: &str)
                 .collect();
 
             let ref_config = ReferenceConfig {
-                quad_decimate: s.quad_decimate.unwrap_or(2.0) as f32,
+                quad_decimate: s.quad_decimate.unwrap_or(2.0),
                 ..Default::default()
             };
